requirements/dev.txt
django-debug-toolbar==1.3.2
django-nose==1.4.1
factory-boy==2.5.2
nose==1.3.7
pip-tools
